+++ src/EventSourcing.EventStoreDB/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("ShopSharp.Users.EventSourcing.EventStoreDB.IntegrationTests")]